+++ rust-src/concordium_base/src/protocol_level_tokens/token_event.rs
@@ -1,5 +1,8 @@
 use crate::{
-    common::{cbor, cbor::CborSerializationResult},
+    common::cbor::{
+        self, CborDeserialize, CborMapDecoder, CborSerializationError, CborSerializationResult,
+        CborSerialize,
+    },
     transactions::Memo,
 };
 use concordium_base_derive::{CborDeserialize, CborSerialize};
@@ -38,6 +41,7 @@ pub enum TokenEventDetails {
 #[serde(rename_all = "camelCase")]
 pub struct TokenModuleEvent {
     /// The type of event produced.
+    #[serde(rename = "type")]
     pub event_type: TokenModuleCborTypeDiscriminator,
     /// The details of the event produced, in the raw byte encoded form.
     pub details:    RawCbor,
@@ -53,6 +57,8 @@ impl TokenModuleEvent {
             "removeAllowList" => RemoveAllowList(cbor::cbor_decode(self.details.as_ref())?),
             "addDenyList" => AddDenyList(cbor::cbor_decode(self.details.as_ref())?),
             "removeDenyList" => RemoveDenyList(cbor::cbor_decode(self.details.as_ref())?),
+            "pause" => Pause(cbor::cbor_decode(self.details.as_ref())?),
+            "unpause" => Unpause(cbor::cbor_decode(self.details.as_ref())?),
             _ => Unknow,
         })
     }
@@ -70,6 +76,12 @@ pub enum TokenModuleEventType {
     AddDenyList(TokenListUpdateEventDetails),
     /// An account was removed from the deny list of a protocol level token
     RemoveDenyList(TokenListUpdateEventDetails),
+    /// Execution of certain operations on a protocol level token was
+    /// paused
+    Pause(TokenPauseEventDetails),
+    /// Execution of certain operations on a protocol level token was
+    /// unpaused
+    Unpause(TokenPauseEventDetails),
     /// Unknow token module event type. If new events types are added that are
     /// unknown to this enum, they will be decoded to this variant.
     Unknow,
@@ -91,6 +103,30 @@ pub enum TokenModuleEventType {
 pub struct TokenListUpdateEventDetails {
     /// The account that was added or removed from an allow or deny list
     pub target: CborTokenHolder,
+}
+
+/// An event emitted when the token is paused or unpaused.
+#[derive(Debug, Clone, Eq, PartialEq, serde::Serialize, serde::Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct TokenPauseEventDetails {}
+
+impl CborSerialize for TokenPauseEventDetails {
+    fn serialize<C: cbor::CborEncoder>(&self, encoder: C) -> CborSerializationResult<()> {
+        encoder.encode_map(0)?;
+        Ok(())
+    }
+}
+
+impl CborDeserialize for TokenPauseEventDetails {
+    fn deserialize<C: cbor::CborDecoder>(decoder: C) -> CborSerializationResult<Self>
+    where
+        Self: Sized, {
+        let map_decoder = decoder.decode_map()?;
+        if map_decoder.size() != 0 {
+            return Err(CborSerializationError::map_size(0, map_decoder.size()));
+        }
+        Ok(TokenPauseEventDetails {})
+    }
 }
 
 /// An entity that can hold PLTs (protocol level tokens).
@@ -282,5 +318,33 @@ mod test {
             module_event_type,
             TokenModuleEventType::RemoveDenyList(variant)
         );
+    }
+
+    #[test]
+    fn test_decode_pause_event_cbor() {
+        let variant = TokenPauseEventDetails {};
+        let cbor = cbor::cbor_encode(&variant).unwrap();
+        assert_eq!(hex::encode(&cbor), "a0");
+        let module_event = TokenModuleEvent {
+            event_type: "pause".to_string().try_into().unwrap(),
+            details:    cbor.into(),
+        };
+
+        let module_event_type = module_event.decode_token_module_event().unwrap();
+        assert_eq!(module_event_type, TokenModuleEventType::Pause(variant));
+    }
+
+    #[test]
+    fn test_decode_unpause_event_cbor() {
+        let variant = TokenPauseEventDetails {};
+        let cbor = cbor::cbor_encode(&variant).unwrap();
+        assert_eq!(hex::encode(&cbor), "a0");
+        let module_event = TokenModuleEvent {
+            event_type: "unpause".to_string().try_into().unwrap(),
+            details:    cbor.into(),
+        };
+
+        let module_event_type = module_event.decode_token_module_event().unwrap();
+        assert_eq!(module_event_type, TokenModuleEventType::Unpause(variant));
     }
 }